- typecast error fixed

+++ code/go/0chain.net/smartcontract/stakepool/edb_delegate_pool.go
@@ -26,7 +26,7 @@ func (dp DelegatePool) emitNew(
 	balances cstate.StateContextI,
 ) error {
 	data, err := json.Marshal(&event.DelegatePool{
-		Balance:      int64(dp.Balance),
+		Balance:      dp.Balance,
 		PoolID:       poolId,
 		ProviderType: int(providerType),
 		ProviderID:   providerId,
